[Mono.Android] Share string-based JNI native registration
Extract the string-based JNI native method registration logic from
AndroidTypeManager.RegisterNativeMembers into a shared
NativeMethodRegistration helper. The helper owns parsing the method
metadata string, resolving callback delegates, handling [Export] dynamic
callbacks, rooting those callbacks, and registering natives with JNI.

Use the helper from AndroidTypeManager after the linker-generated fast
registration map, preserving the existing fallback to Java.Interop's
marshal-method registration when applicable. Also use the same helper from
TrimmableTypeMapTypeManager when the new string-based registration feature
switch is enabled.

Rename the feature to RuntimeFeature.StringBasedJniRegistration and guard
it with [FeatureGuard(typeof(RequiresUnreferencedCodeAttribute))], because
this path relies on reflection over method metadata strings. The switch is
true by default for non-trimmable typemaps and false by default for the
trimmable typemap, where native registration normally happens through
mono.android.Runtime.registerNatives(Class).

Move the trimmable registerNatives JNI callback wiring out of
TrimmableTypeMap and into JNIEnvInit, so responsibilities are clearer:
JNIEnvInit wires JNI callbacks, TrimmableTypeMap resolves generated
wrappers, and TrimmableTypeMapTypeManager handles JniTypeManager behavior.
